+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/FlatsExpert.cs
@@ -525,7 +525,7 @@ namespace Assistant.NINAPlugin.Sequencer {
             Offset = offset;
             BinningMode = binning;
             ReadoutMode = readoutMode;
-            Rotation = rotation;
+            Rotation = RotationZero(rotation);
             ROI = roi;
             Key = GetKey();
         }
@@ -539,7 +539,7 @@ namespace Assistant.NINAPlugin.Sequencer {
             BinningMode.TryParse(exposure.Metadata.Binning, out bin);
             BinningMode = bin;
             ReadoutMode = exposure.Metadata.ReadoutMode;
-            Rotation = exposure.Metadata.RotatorMechanicalPosition;
+            Rotation = RotationZero(exposure.Metadata.RotatorMechanicalPosition);
             ROI = exposure.Metadata.ROI;
             Key = GetKey();
         }
@@ -551,19 +551,34 @@ namespace Assistant.NINAPlugin.Sequencer {
             Offset = flatHistory.Offset;
             BinningMode = flatHistory.BinningMode;
             ReadoutMode = flatHistory.ReadoutMode;
-            Rotation = flatHistory.Rotation;
+            Rotation = RotationZero(flatHistory.Rotation);
             ROI = flatHistory.ROI;
             Key = GetKey();
         }
 
         private string GetKey() {
-            return $"{TargetId}_{FilterName}_{Gain}_{Offset}_{BinningMode}_{ReadoutMode}_{ROI}";
+            return $"{FilterName}_{Gain}_{Offset}_{BinningMode}_{ReadoutMode}_{ROI}";
+        }
+
+        private double RotationZero(double rotation) {
+            if (rotation == ImageMetadata.NO_ROTATOR_ANGLE || Double.IsNaN(rotation)) {
+                return 0.0;
+            }
+
+            return rotation;
         }
 
         public bool Equals(FlatSpec other) {
             if (other is null) { return false; }
             if (ReferenceEquals(this, other)) { return true; }
             if (GetType() != other.GetType()) { return false; }
+
+            // If rotation isn't a factor for these targets OR it's the same target, then just compare on key
+            if (Rotation == 0.0 && other.Rotation == 0.0) { return Key == other.Key; }
+            if (TargetId == other.TargetId) { return Key == other.Key; }
+
+            // Targets are different - if rotation is different then we're not the same
+            if (Rotation != other.Rotation) { return false; }
 
             return Key == other.Key;
         }